% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.ttest.R
\name{multi.ttest}
\alias{multi.ttest}
\title{多重t-test计算
这个函数用于计算多个变量，两两之间的t-test}
\usage{
multi.ttest(mat, ...)
}
\arguments{
\item{mat}{dataframe}

\item{...}{}
}
\value{

}
\description{
多重t-test计算
这个函数用于计算多个变量，两两之间的t-test
}
