% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmatch_analysis.R
\name{pmatch_analysis}
\alias{pmatch_analysis}
\title{批量匹配函数
先调用pmatch_analysis函数，再根据以下的代码，进入循环。}
\usage{
pmatch_analysis(query, target)
}
\arguments{
\item{query}{要查询的ID列}

\item{target}{要匹配的数据库}
}
\value{

}
\description{
批量匹配函数
先调用pmatch_analysis函数，再根据以下的代码，进入循环。
}
