% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD_summary.R
\name{SD_summary}
\alias{SD_summary}
\title{标准差汇总
该程序用于计算数据框的平均值和标准差SD}
\usage{
SD_summary(data, varname, groupnames, digtal)
}
\arguments{
\item{data}{数据框}

\item{varname}{目标数据列}

\item{groupnames}{分组}

\item{digtal}{结果的小数点}
}
\value{

}
\description{
标准差汇总
该程序用于计算数据框的平均值和标准差SD
}
\examples{
SD_summary(iris,varname = "Petal.Width", groupnames = "Species", digtal = 2)
}
