% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_summary.R
\name{SE_summary}
\alias{SE_summary}
\title{标准误汇总
该程序用于计算数据框的平均值和标准误SE}
\usage{
SE_summary(data, varname, groupnames, digtal)
}
\arguments{
\item{data}{数据框}

\item{varname}{目标数据列}

\item{groupnames}{分组}

\item{digtal}{结果的小数点}
}
\value{

}
\description{
标准误汇总
该程序用于计算数据框的平均值和标准误SE
}
\examples{
SE_summary(iris,varname = "Petal.Width", groupnames = "Species", digtal = 2)
}
